ルーティングとコントローラの修正
