+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   root to: 'items#index'
-  resources :items do
-    resources :purchase_histories, only: :index
+  resources :orders, only: :index do
+    resources :purchase_historyories, only: :create
   end 
   devise_for :users
 end
